Image Classification/CGIAR Computer Vision for Crop Disease/crop disease/requirements.txt
hyperopt==0.2.3
pandas==1.0.1
torch==1.4.0
scipy==1.4.1
torchvision==0.5.0
numpy==1.18.1
Pillow==7.0.0
scikit_learn==0.22.2.post1
